adding desktop test

+++ test/home.js
@@ -12,6 +12,14 @@ const fs = require('fs-extra');
 let browser;
 let page;
 
+const looksSameOpts = {
+	highlightColor: '#ff00ff',
+	ignoreAntialiasing: true,
+	ignoreCaret: true,
+	//strict: true,
+	tolerance: 2.3,
+};
+
 describe('the home page', () => {
 	before(async () => {
 		browser = await puppeteer.launch({
@@ -25,34 +33,25 @@ describe('the home page', () => {
 	it('looks right on mobile', async () => {
 		page = await browser.newPage();
 		await page.setViewport({
-			width: 360,
-			height: 640,
-			deviceScaleFactor: 2,
-			isMobile: true,
+			deviceScaleFactor: 4,
 			hasTouch: true,
+			height: 640,
 			isLandscape: false,
+			isMobile: true,
+			width: 360,
 		});
 		await page.goto('http://localhost:1234');
-		const currPath = 'shots/home.curr.png';
-		const refPath = 'shots/home.ref.png';
-		const diffPath = 'shots/home.diff.png';
+		const currPath = 'shots/home.mobile.curr.png';
+		const refPath = 'shots/home.mobile.ref.png';
+		const diffPath = 'shots/home.mobile.diff.png';
 		await page.screenshot({
 			path: currPath,
 			fullPage: true,
 		});
 		const oldFileExists = await fs.pathExists(refPath);
 		if (oldFileExists) {
-			console.info("comparing to previous screenshot");
-			const looksSameOpts = {
-				current: currPath,
-				diff: diffPath,
-				highlightColor: '#ff00ff',
-				ignoreAntialiasing: true,
-				ignoreCaret: true,
-				reference: refPath,
-				//strict: true,
-				tolerance: 2.3,
-			};
+			console.info('comparing to previous screenshot');
+
 			const eql = await new Promise((resolve) => {
 				looksSame(currPath, refPath, looksSameOpts, (err, eql) => {
 					if (err) console.error(err);
@@ -60,17 +59,85 @@ describe('the home page', () => {
 				});
 			});
 			if (!eql) {
-				console.info("creating diff image");
+				console.info('creating diff image');
 				await new Promise((resolve) => {
-					looksSame.createDiff(looksSameOpts, (err) => {
-						if (err) console.error(err);
-						resolve();
-					});
+					looksSame.createDiff(
+						Object.assign(
+							{
+								current: currPath,
+								diff: diffPath,
+								reference: refPath,
+							},
+							looksSameOpts,
+						),
+						(err) => {
+							if (err) console.error(err);
+							resolve();
+						},
+					);
 				});
 			}
 			expect(eql, 'return value of looks-same').to.be.true();
 		} else {
-			console.info("saving current screenshot for next time");
+			console.info('saving current screenshot for next time');
+			await fs.move(currPath, refPath);
+			expect(
+				oldFileExists,
+				'a previous screenshot was not found, current screenshot has been saved',
+			).to.be.true();
+		}
+	});
+
+	it('looks right on desktop', async () => {
+		page = await browser.newPage();
+		await page.setViewport({
+			deviceScaleFactor: 3,
+			hasTouch: false,
+			height: 768,
+			isLandscape: true,
+			isMobile: false,
+			width: 1366,
+		});
+		await page.goto('http://localhost:1234');
+		const currPath = 'shots/home.desktop.curr.png';
+		const refPath = 'shots/home.desktop.ref.png';
+		const diffPath = 'shots/home.desktop.diff.png';
+		await page.screenshot({
+			path: currPath,
+			fullPage: true,
+		});
+		const oldFileExists = await fs.pathExists(refPath);
+		if (oldFileExists) {
+			console.info('comparing to previous screenshot');
+
+			const eql = await new Promise((resolve) => {
+				looksSame(currPath, refPath, looksSameOpts, (err, eql) => {
+					if (err) console.error(err);
+					resolve(eql);
+				});
+			});
+			if (!eql) {
+				console.info('creating diff image');
+				await new Promise((resolve) => {
+					looksSame.createDiff(
+						Object.assign(
+							{
+								current: currPath,
+								diff: diffPath,
+								reference: refPath,
+							},
+							looksSameOpts,
+						),
+						(err) => {
+							if (err) console.error(err);
+							resolve();
+						},
+					);
+				});
+			}
+			expect(eql, 'return value of looks-same').to.be.true();
+		} else {
+			console.info('saving current screenshot for next time');
 			await fs.move(currPath, refPath);
 			expect(
 				oldFileExists,
